refactor(admin dashboard): move components to match folder structure and change drawer style
Moved components so that it matches workspace convention. Change drawer to not use MUI Drawer and
make sure to use NextJS router

Clean stylesheet for admin drawer
