+++ components/admin/AdminIndex/AdminIndex.module.css
@@ -9,7 +9,6 @@
   display: flex;
   padding: 20px;
   padding-top: 0px;
-  margin-left: 180px;
 }
 
 .admin {
@@ -22,7 +21,6 @@
   display: flex;
   flex-direction: row;
   width: 100%;
-  margin-top: 40px;
 }
 
 .leftCol {
@@ -32,7 +30,7 @@
 }
 
 .rightCol {
-  width: 100%;
+  width: calc(100% - 200px);
 }
 
 .buttons {
