+++ styles/Layout.module.css
@@ -1,9 +1,10 @@
-.pageTitle {
+.navbar {
   display: flex;
   justify-content: space-between;
   align-items: center;
   padding-left: 25px;
   padding-right: 25px;
+  margin-bottom: 20px;
 }
 
 .logButtonSpace {
