--- styles/admin/AdminDrawer.module.css
@@ -1,32 +1,0 @@
-.drawer {
-  margin-top: 100px;
-  flex-shrink: 0;
-  width: 240;
-}
-.root {
-  display: flex;
-}
-.appBar {
-  margin-top: 90px;
-  width: 180px;
-}
-.title {
-  display: flex;
-  font-weight: 800;
-  font-size: 50px;
-  width: 150px;
-  padding-left: 20px;
-}
-
-.selectedItemStyle {
-  background-color: rgb(0, 156, 143);
-}
-
-.buttonFlex {
-  margin-left: 8px;
-  margin-bottom: 10px;
-}
-
-.selected {
-  margin-bottom: 10px;
-}
